chore(kreivo-runtime): finish splitting runtime config
