--- runtime/kreivo/src/governance/mod.rs
@@ -1,3 +1,0 @@
-pub mod origins;
-
-pub use origins::*;
